Create x509 certificate and add it to http client
First it attemps to load a certificate from the local storage.
If it does not find a certificate file, it attemps to generate a new
private/public key and a certificate from it. It then saves the
certificate, along with the private key, to a file in the PEM format.
This certificate is added to the http client for the TLS connection.